chore: move file to better location
